requirements.txt
pyrogram>=2.0.106
tgcrypto
pymongo[srv]>=4.6.0
python-dotenv
flask
certifi